requirements.txt
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyNaCl==1.4.0
six==1.16.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
